+++ tests/endtoend/opart/comparison/emispure/forward_time_opart.py
@@ -1,5 +1,5 @@
 from exojax.opacity import OpaPremodit
-from exojax.rt.opart import OpartEmisPure
+from exojax.rt import OpartEmisPure
 from exojax.rt.layeropacity import single_layer_optical_depth
 from exojax.utils.grids import wavenumber_grid
 from exojax.database.api  import MdbExomol
