+++ tests/endtoend/opart/comparison/reflectpure/forward_reflect_opart.py
@@ -1,5 +1,5 @@
 from exojax.opacity import OpaPremodit
-from exojax.rt.opart import OpartReflectPure
+from exojax.rt import OpartReflectPure
 from exojax.rt.layeropacity import single_layer_optical_depth
 from exojax.utils.grids import wavenumber_grid
 from exojax.database.api  import MdbExomol
